scripts/firestore-setup.py
import firebase_admin
from firebase_admin import credentials
from firebase_admin import firestore

# Use the application default credentials
cred = credentials.ApplicationDefault()
firebase_admin.initialize_app(cred, {"projectId": "zbb-239719"})

db = firestore.client()
