requirements.txt
fastapi==0.20.0
google-cloud-firestore==1.1.0
starlette==0.11.4
uvicorn==0.7.1
